Updated starter project to latest version with NextJS
