+++ app/pages/index.js
@@ -1,0 +1,24 @@
+import React from "react";
+
+// Constants
+const TWITTER_HANDLE = "_buildspace";
+const TWITTER_LINK = `https://twitter.com/${TWITTER_HANDLE}`;
+
+const Home = () => {
+    return (
+        <div className="App">
+            <div className="container">
+                <div className="header-container">
+                    <p className="header">🍭 Candy Drop</p>
+                    <p className="sub-text">NFT drop machine with fair mint</p>
+                </div>
+                <div className="footer-container">
+                    <img alt="Twitter Logo" className="twitter-logo" src={twitterLogo} />
+                    <a className="footer-text" href={TWITTER_LINK} target="_blank" rel="noreferrer">{`built on @${TWITTER_HANDLE}`}</a>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default Home;
