--- app/src/CandyMachine/CandyMachine.css
@@ -1,33 +1,0 @@
-.machine-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.gif-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.gif-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  grid-gap: 1.5rem;
-  justify-items: center;
-  margin: 0;
-  padding: 20px 0 20px 0;
-}
-
-.gif-grid .gif-item {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  justify-self: center;
-  align-self: center;
-}
-
-.gif-item img {
-  width: 100%;
-  height: 200px;
-  border-radius: 10px;
-  object-fit: cover;
-}
