--- app/src/CountdownTimer/CountdownTimer.css
@@ -1,8 +1,0 @@
-.timer-container .timer-header {
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.timer-container .timer-value {
-  font-size: 18px;
-}
